I implemented a few optimizations that increase performance by about 200%

<table>
<tr><td>Before (original code)</td><td>After</td></tr>
<tr><td>

```
Running 498 test cases...
compiler: MSVC v19.34.31937
architecture: x64
__cpp_lib_string_view: 201803
sizeof wchar_t: 2
UTFW: UTF16
Resolution: 1497604738
UTF8  ==> UTF8 : 0.046760979s
UTF8  ==> UTF16: 0.120738046s
UTF8  ==> UTF32: 0.116526514s
UTF8  ==> UTFW : 0.125314060s
UTF16 ==> UTF8 : 0.149225085s
UTF16 ==> UTF16: 0.025468075s
UTF16 ==> UTF32: 0.044913877s
UTF16 ==> UTFW : 0.037495761s
UTF32 ==> UTF8 : 0.149762023s
UTF32 ==> UTF16: 0.080844693s
UTF32 ==> UTF32: 0.016662613s
UTF32 ==> UTFW : 0.083288181s
UTFW  ==> UTF8 : 0.148268501s
UTFW  ==> UTF16: 0.039871186s
UTFW  ==> UTF32: 0.046740020s
UTFW  ==> UTFW : 0.028833317s
codecvt_utf8_utf16<char16_t>:
UTF16 ==> UTF8 : 0.312078935s (+109.13%)
UTF8  ==> UTF16: 0.229479163s (+90.06%)
codecvt_utf8_utf16<wchar_t>:
UTFW  ==> UTF8 : 0.325835292s (+119.76%)
UTF8  ==> UTFW : 0.235868838s (+88.22%)

*** No errors detected
```

</td><td>

```
Running 532 test cases...
compiler: MSVC v19.34.31937
architecture: x64
__cpp_lib_string_view: 201803
sizeof wchar_t: 2
UTFW: UTF16
Resolution: 1497608346
UTF8  ==> UTF8 : 0.020368061s
UTF8  ==> UTF16: 0.054659417s
UTF8  ==> UTF32: 0.045423582s
UTF8  ==> UTFW : 0.052922389s
UTF16 ==> UTF8 : 0.051900870s
UTF16 ==> UTF16: 0.011311742s
UTF16 ==> UTF32: 0.031817055s
UTF16 ==> UTFW : 0.011952694s
UTF32 ==> UTF8 : 0.041485858s
UTF32 ==> UTF16: 0.032115685s
UTF32 ==> UTF32: 0.009692607s
UTF32 ==> UTFW : 0.033776585s
UTFW  ==> UTF8 : 0.047951779s
UTFW  ==> UTF16: 0.013337822s
UTFW  ==> UTF32: 0.031452300s
UTFW  ==> UTFW : 0.012163787s
codecvt_utf8_utf16<char16_t>:
UTF16 ==> UTF8 : 0.284035091s (+447.26%)
UTF8  ==> UTF16: 0.191992105s (+251.25%)
codecvt_utf8_utf16<wchar_t>:
UTFW  ==> UTF8 : 0.285054652s (+494.46%)
UTF8  ==> UTFW : 0.202671043s (+282.96%)

*** No errors detected
```

</td>
</tr>
</table>

To get to this result I implemented serveral changes:

- Changed `run_measure()` in the performance test from using `std::back_inserter()` to use a pointer to the start of the output buffer. Having `back_inserter()` as the output iterator pollutes the measurment by adding a part of `std::vector` into the test.
- Added force-inline attribute to all `read()` and `write()` methods. UTF-8 `read()` was too large to qualify for inlining.

One major change is in how `conv_strategy` references the input the output iterators.  
The original code captured the input iterator by reference in a lambda. This caused the code that was generated in `read()` to have to do 2 indirections:

```           
char_type const ch1 = read_fn(); if (ch1 >> 6 != 2) goto _err;
         mov         rcx,qword ptr [rcx]  
         mov         rax,qword ptr [rcx]  
         movzx       edx,byte ptr [rax]  
         inc         rax  
         mov         qword ptr [rcx],rax  
```
In  the new code this is changed to:
```            
char_type const ch1 = read_fn.read1(); 
         movzx       ecx,byte ptr [rax]  
         inc         rax  
         mov         qword ptr [r8],rax  
```

Fixing this was the motivating for replacing the lambdas with a set of iterator wrapper classes that takes the input iterator by value (`ReadIterRnd`). Having an iterator wrapper class allowed a few more optimizations to be subsequently added:

- Instead of reading input elements one by one, read them a few at a time. This is significant in the UTF-8 reader that can read 3-4 bytes at a time, and leads to better generated code.
- An explicit specialization for the `const char*` reader case was added where input bytes are converts to `uint32_t` before returning to `read()`. This allows better generated code since the optimizer can deal with full sized registers instead of bytes.
- `ReadIterRndChecked` is used for the last few code points that require bounds checking. the check is made per group of 2-5 elements instead of on every element (byte for UTF-8) read.
- UTF-8 `write()` was changed to do with less brancing (at the cost of duplicating code using macros)

A few missing test-cases were added

- Converting a strings that end in a partial code-point should throw an exception
- Usage of non-random-access iterators

Tested with Visual Studio 2022, toolkit v143.  
The code uses a few modern features like std::tuple and structured binding so it doesn't compile on older toolkits. If this support is a priority I can try to rework these parts.

Future work:
- It would be interesting to try to implement UTF-8 recovery and resync after detecting a bad sequence in the middle of a string, and do so without a performance impact to the normal case.

------
The goal of this PR was to try to get your attention to my recent job application to the Profiler Core team at JetBrains. I thought seeing code could be better than any cover letter I could write.  
If you're looking for someone who can read and understand code fast, enjoys tinkering with assembly code and working for 3 days to make the optimizer do a better job, please contact me - shooshx@gmail.com